Crate foxbox_taxonomy

This crate defines the high-level API for accessing Connected Devices.

Modules

adapter
The API for defining Adapters.

adapter_utils
Utilities for writing Adapters. Utilities for writing adapters.

api
Public-facing API

fake_adapter
Implementation of a fake adapter, controlled entirely programmatically. Designed to be used as a component of tests. An adapter to a non-existing device, whose state is entirely controlled programmatically. Used for testing.

manager
The manager provides an API for (un)registering adapters, services, channels, and uses these to implements the taxonomy API. The Adapter manager

parse
Tools for parsing from JSON. Utilities for defining a JSON parser.

selector
Selecting one or more devices. Exposed through the API. Selectors for services and channels.

serialize
Tools for converting values to a format that may be sent. Utilities for serializing data to JSON.

services
Metadata on devices This module defines the metadata on devices and services.

transact
Utility module for inserting values in maps and keeping the insertion reversible in case of any error. Implementation of reversible insertions on maps.

util
Various utilities

values
Values that may be sent to/received from devices

Macros

adapter_id!

service_id!

tag_id!