Module foxbox_taxonomy::api

Public-facing API

The API for communicating with devices.

This API is provided as Traits to be implemented:

by the low-level layers of the FoxBox, including the adapters;

by test suites and tools that need to simulate connected devices.

In turn, this API is used to implement:

the public-facing REST and WebSocket API;

the rules API (ThinkerBell).

Reexports

pub use util::{ResultMap, TargetMap, Targetted};

Enums

Error
An error that arose during interaction with either a device, an adapter or the adapter manager

InternalError

User
User identifier that will be passed from the REST API handlers to the adapters.

WatchEvent
An event during watching.

Traits

API
A handle to the public API.