Module foxbox_taxonomy::devices

Metadata on devices This module defines the metadata on devices and services.

Note that all the data structures in this module represent snapshots of subsets of the devices available. None of these data structures are live, so there is always the possibility that devices may have been added or removed from the FoxBox by the time these data structures are read.

Structs

Channel
An channel represents a single place where data can enter or leave a device. Note that channels support either a single kind of getter or a single kind of setter. Devices that support both getters or setters, or several kinds of getters, or several kinds of setters, are represented as nodes containing several channels.

Getter
A getter operation available on a channel.

Node
Metadata on a node. A node is a device or collection of devices that may offer services. The FoxBox itself a node offering services such as a clock, communication with the user through her smart devices, etc.

NodeId
A marker for Id. Only useful for writing Id<NodeId>.

Setter
An setter operation available on an channel.

Enums

ChannelKind
The kind of the service, i.e. a strongly-typed description of what the service can do. Used both for locating services (e.g. "I need a clock" or "I need something that can provide pictures") and for determining the data structure that these services can provide or consume.

Traits

IOMechanism
The communication mechanism used by the channel.