Enum foxbox_taxonomy::fake_adapter::Tweak

pub enum Tweak {
    InjectGetterValue(Id<Getter>, Result<Option<Value>, Error>),
    InjectSetterError(Id<Setter>, Option<Error>),
}

A tweak sent to the virtual device, to set a value, inject an error, ...

Variants

InjectGetterValue

Inject a value in a virtual getter.

InjectSetterError

Inject an error in a virtual setter. All operations on this setter will raise the error until None is injected instead.