Struct foxbox_taxonomy::manager::AdapterManager

pub struct AdapterManager {
    // some fields omitted
}

An implementation of the AdapterManager.

This implementation is Sync and supports any number of concurrent readers or a single writer.

Methods

impl AdapterManager

fn new() -> Self

Create an empty AdapterManager. This function does not attempt to load any state from the disk.