Module foxbox_taxonomy::selector

Selecting one or more devices. Exposed through the API. Selectors for services and channels.

The high-level API of Project Link always offers access by selectors, rather than by individual services/channels. This allows operations such as sending a temperature to all heaters in the living room (that's a selector), rather than needing to access every single heater one by one.

Reexports

pub use parse::*;

Structs

GetterSelector
A selector for one or more getter channels.

Period
An acceptable interval of time.

ServiceSelector
A selector for one or more services.

SetterSelector
A selector for one or more setter channels.

Traits

SelectedBy

ServiceLike
A trait used to let ServiceSelector work on complex data structures that are not necessarily exactly Selector.