Module foxbox_taxonomy::services

Metadata on devices This module defines the metadata on devices and services.

Note that all the data structures in this module represent snapshots of subsets of the devices available. None of these data structures are live, so there is always the possibility that devices may have been added or removed from the FoxBox by the time these data structures are read.

Reexports

pub use util::{Exactly, Id, AdapterId, ServiceId, KindId, TagId, VendorId};

Structs

Channel
An channel represents a single place where data can enter or leave a device. Note that channels support either a single kind of getter or a single kind of setter. Devices that support both getters or setters, or several kinds of getters, or several kinds of setters, are represented as services containing several channels.

Getter
A getter operation available on a channel.

Service
Metadata on a service. A service is a device or collection of devices that may offer services. The FoxBox itself is a service offering services such as a clock, communicating with the user through her smart devices, etc.

Setter
An setter operation available on an channel.

Enums

ChannelKind
The kind of the channel, i.e. a strongly-typed description of what the channel can do. Used both for locating channels (e.g. "I need a clock" or "I need something that can provide pictures") and for determining the data structure that these channel can provide or consume.

Traits

IOMechanism
The communication mechanism used by the channel.