Struct foxbox_taxonomy::services::Channel

pub struct Channel<IO> where IO: IOMechanism {
    pub tags: HashSet<Id<TagId>>,
    pub id: Id<IO>,
    pub service: Id<ServiceId>,
    pub mechanism: IO,
    pub adapter: Id<AdapterId>,
    pub last_seen: Option<TimeStamp>,
}

An channel represents a single place where data can enter or leave a device. Note that channels support either a single kind of getter or a single kind of setter. Devices that support both getters or setters, or several kinds of getters, or several kinds of setters, are represented as services containing several channels.

Fields

tags

Tags describing the channel.

These tags can be set by the user, adapters or applications. They are used to regroup channels for rules.

For instance "entrance".

id

An id unique to this channel.

service

The service owning this channel.

mechanism

The update mechanism for this channel.

adapter

Identifier of the adapter for this channel.

last_seen

The last time the device was seen.