Struct foxbox_taxonomy::services::Service

pub struct Service {
    pub tags: HashSet<String>,
    pub id: Id<ServiceId>,
    pub getters: HashMap<Id<Getter>, Channel<Getter>>,
    pub setters: HashMap<Id<Setter>, Channel<Setter>>,
    pub adapter: Id<AdapterId>,
}

Metadata on a service. A service is a device or collection of devices that may offer services. The FoxBox itself is a service offering services such as a clock, communicating with the user through her smart devices, etc.

Fields

tags

Tags describing the service.

These tags can be set by the user, adapters or applications. They are used by applications to find services and services.

For instance, a user may set tag "entrance" to all services placed in the entrance of his house, or a tag "blue" to a service controlling blue lights. An adapter may set tags "plugged" or "battery" to devices that respectively depend on a plugged power source or on a battery.

id

An id unique to this service.

getters

Channels connected directly to this service.

setters

adapter

Identifier of the adapter for this service.

Methods

impl Service

fn empty(id: Id<ServiceId>, adapter: Id<AdapterId>) -> Self

Create an empty service.