Module foxbox_taxonomy::transact

Utility module for inserting values in maps and keeping the insertion reversible in case of any error. Implementation of reversible insertions on maps.

These utility data structures are useful when several hashmaps/hashsets need to be kept synchronized. For instance, maps a data structure needs to be added to maps a, b, c but the entire operation needs to be cancelled if there is a collision in map b.

Structs

InsertInMap
Insert a (key, value) pair in a map. However, if the object is dropped before method commit() is called, the insertion is cancelled.