Module foxbox_taxonomy::util

Various utilities

Structs

AdapterId
A marker for Id. Only useful for writing Id<AdapterId>.

Id
A unique id for values of a given kind.

KindId

MimeTypeId

Phantom
A variant of PhantomData that supports [De]serialization

ServiceId
A marker for Id. Only useful for writing Id<ServiceId>.

TagId
Only useful for writing Id<TagId>.

Targetted

TrivialEnumVisitor
By default, the (de)serialization of trivial enums by Serde is surprising, e.g. in JSON, enum Foo {A, B, C} will produce {"\"A\": []"} for A, where "\"A\"" would be expected.

VendorId

Enums

Exactly
A marker for a request that a expects a specific value.

Functions

is_sync
Helper function, to check that a type implements Sync.

Type Definitions

ResultMap
A bunch of results coming from different sources.

TargetMap
A bunch of instructions, going to different targets.