Module foxbox_taxonomy::values

Values that may be sent to/received from devices

Values manipulated by services

Structs

Binary

Duration
A duration, also used to represent a time of day.

ExtValue
A data structure holding a boolean value of a type that has not been standardized yet.

Json
Representation of an object in JSON. It is often (albeit not always) possible to choose a more precise data structure for representing values send/accepted by a service. If possible, adapters should rather pick such more precise data structure.

ThinkerbellRule

TimeStamp
An absolute time and date.

TypeError
Representation of a type error.

Enums

Color
A color. Internal representation may vary. The FoxBox adapters are expected to perform conversions to the format requested by their device.

OnOff
An on/off state.

OpenClosed
An open/closed state.

Range
A comparison between two values.

Temperature
A temperature. Internal representation may be either Fahrenheit or Celcius. The FoxBox adapters are expected to perform conversions to the format requested by their devices.

Type
The type of values manipulated by endpoints.

Value
Representation of an actual value that can be sent to/received from a service.

Implementations