Struct foxbox_taxonomy::values::ExtValue

pub struct ExtValue<T> where T: Debug + Clone + PartialEq + PartialOrd + Serialize + Deserialize {
    pub value: T,
    pub vendor: Id<VendorId>,
    pub adapter: Id<AdapterId>,
    pub kind: Id<KindId>,
}

A data structure holding a boolean value of a type that has not been standardized yet.

Fields

value

vendor

The vendor. Used for namespacing purposes, to avoid confusing two incompatible extensions with similar names. For instance, "foxlink@mozilla.com".

adapter

Identification of the adapter introducing this value. Designed to aid with tracing and debugging.

kind

A string describing the nature of the value, designed to aid with type-checking.

Examples: "PresenceDetected".